Updated all ColliderBit analyses to use event->weight() when incrementing signal region counters. *phew*
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_0LEPStop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_0LEP_13invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_0LEP_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_1LEPStop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_2LEPStop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_2OSLEP_Z_139invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_2OSLEP_chargino_139invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_2OSLEP_chargino_80invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_2bMET_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_3b_24invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_3b_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_4LEP_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_MultiLEP_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_MultiLEP_confnote_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_PhotonGGM_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_RJ3L_lowmass_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_13TeV_ZGammaGrav_CONFNOTE_80invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_7TeV_1OR2LEPStop_4_7invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_7TeV_2LEPStop_4_7invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_0LEPStop_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_0LEP_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_1LEPStop_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_1LEPbb_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_2LEPEW_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_2LEPStop_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_2bStop_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_ATLAS_8TeV_3LEPEW_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_0LEP_13invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_0LEP_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_0LEP_36invfb_withCovariance.cpp.inactive
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_1LEPStop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_1LEPbb_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2LEPStop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2LEPsoft_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2OSLEP_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2OSLEP_chargino_stop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2OSLEP_confnote_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2Photon_GMSB_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2SSLEP_Stop_137invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_2SSLEP_Stop_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_MONOJET_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_MultiLEP_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_MultiLEP_Full_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_13TeV_Photon_GMSB_36invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_8TeV_1LEPDMTOP_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_8TeV_2LEPDMTOP_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_8TeV_MONOJET_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_CMS_8TeV_MultiLEP_20invfb.cpp
	modified:   ColliderBit/src/analyses/Analysis_Covariance.cpp
	modified:   ColliderBit/src/analyses/Analysis_Minimum.cpp

+++ ColliderBit/src/analyses/Analysis_Covariance.cpp
@@ -12,7 +12,7 @@ namespace Gambit {
 
       // Variables that holds the number of events passing
       // signal region cuts
-      int _numSR;
+      double _numSR;
 
     public:
